[REFACTOR, TEST] Menu 리팩토링 및 테스트 #98

+++ src/main/java/bleuauction/bleuauction_be/server/menu/service/MenuService.java
@@ -1,20 +1,19 @@
 package bleuauction.bleuauction_be.server.menu.service;
 
 import bleuauction.bleuauction_be.server.attach.entity.Attach;
+import bleuauction.bleuauction_be.server.attach.service.AttachService;
 import bleuauction.bleuauction_be.server.menu.entity.Menu;
-import bleuauction.bleuauction_be.server.menu.entity.MenuSize;
 import bleuauction.bleuauction_be.server.menu.entity.MenuStatus;
 import bleuauction.bleuauction_be.server.menu.repository.MenuRepository;
 import bleuauction.bleuauction_be.server.ncp.NcpObjectStorageService;
-import bleuauction.bleuauction_be.server.notice.entity.Notice;
-import bleuauction.bleuauction_be.server.notice.entity.NoticeStatus;
+import bleuauction.bleuauction_be.server.store.entity.Store;
+import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.annotations.CurrentTimestamp;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -23,47 +22,94 @@ import java.util.List;
 public class MenuService {
 
   private final MenuRepository menuRepository;
+  private final AttachService attachService;
   private final NcpObjectStorageService ncpObjectStorageService;
+
 
   //등록
   @Transactional
-  public Long enroll(Menu menu) {
+  public Long enroll(Menu menu, Store store, List<MultipartFile> multipartFiles) {
+
+    menu.setStoreNo(store);
     menuRepository.save(menu);
+
+    if (multipartFiles != null && !multipartFiles.isEmpty()) {
+      for (MultipartFile multipartFile : multipartFiles) {
+        if (multipartFile.getSize() > 0) {
+          Attach attach = ncpObjectStorageService.uploadFile(new Attach(),
+                  "bleuauction-bucket", "menu/", multipartFile);
+          menu.addAttach(attach);
+        }
+      }
+    }
     return menu.getMenuNo();
   }
 
-  //메뉴 전체 조회
-  @Transactional(readOnly = true)
-  public List<Menu> findmenus() {
-    return menuRepository.findAll();
+  public List<Menu> findMenusByStoreNo(Long storeNo) {
+    try {
+      return menuRepository.findMenusByStoreNoAndMenuStatus(storeNo, MenuStatus.Y);
+    } catch (Exception e) {
+      e.printStackTrace();
+      return new ArrayList<>();
+    }
+  }
+
+  public List<Menu> findMenusByStoreNoAndStatus(Long storeNo, MenuStatus menuStatus){
+    try {
+      return menuRepository.findMenusByStoreNoAndMenuStatus(storeNo, menuStatus);
+    } catch (Exception e) {
+      e.printStackTrace();
+      return new ArrayList<>();
+    }
   }
 
 
   //메뉴 1건 조회
-  @Transactional(readOnly = true)
+  @Transactional
   public Menu findOne(Long menuNo) {
-    return menuRepository.findOne(menuNo);
+    return menuRepository.findMenusByMenuNo(menuNo);
   }
 
   //메뉴 삭제(N)
   @Transactional
-  public void deleteMenu(Long menuNo) {
-    Menu menu = menuRepository.findOne(menuNo);
-    menu.delete();
+  public void deleteMenuByMenuNoAndStore(Long menuNo, Store store) {
+    Menu menu = menuRepository.findMenusByMenuNo(menuNo);
+
+    if (menu != null && menu.getStoreNo().equals(store)) {
+      for (Attach attach : menu.getMenuAttaches()) {
+        attachService.changeFileStatusToDeleteByFileNo(attach.getFileNo());
+      }
+      menu.delete();
+    } else {
+      throw new IllegalArgumentException("메뉴와 가게 정보가 유효하지 않습니다.");
+    }
   }
 
   //메뉴 수정
   @Transactional
-  public Menu update(Menu menu) {
-    Menu updatemenu = menuRepository.findOne(menu.getMenuNo());
+  public Menu update(Menu updatedMenu, List<MultipartFile> multipartFiles, Store store) {
+    Menu existingMenu = menuRepository.findMenusByMenuNo(updatedMenu.getMenuNo());
 
-    updatemenu.setMenuName(menu.getMenuName());
-    updatemenu.setMenuSize(menu.getMenuSize());
-    updatemenu.setMenuPrice(menu.getMenuPrice());
-    updatemenu.setMenuContent(menu.getMenuContent());
-    Menu update = menuRepository.save(updatemenu);
+    if (existingMenu.getStoreNo().equals(store)) {
+      existingMenu.setMenuName(updatedMenu.getMenuName());
+      existingMenu.setMenuSize(updatedMenu.getMenuSize());
+      existingMenu.setMenuPrice(updatedMenu.getMenuPrice());
+      existingMenu.setMenuContent(updatedMenu.getMenuContent());
 
-    return updatemenu;
+      if (multipartFiles != null && !multipartFiles.isEmpty()) {
+        for (MultipartFile multipartFile : multipartFiles) {
+          if (multipartFile.getSize() > 0) {
+            Attach attach = ncpObjectStorageService.uploadFile(new Attach(),
+                    "bleuauction-bucket", "menu/", multipartFile);
+            existingMenu.addAttach(attach);
+          }
+        }
+      }
+
+      return menuRepository.save(existingMenu);
+    } else {
+      throw new IllegalArgumentException("수정 권한이 없습니다.");
+    }
   }
